requirements.txt
pylibftdi
python-rtmidi
